requirements.txt
pynecone
cmake
fastapi
opencv-python
dlib
imutils
numpy
python-dotenv
